+++ pumpkin/src/fluid/fluids/mod.rs
@@ -1,2 +1,3 @@
 pub(crate) mod flowing_fluid;
+pub(crate) mod lava;
 pub(crate) mod water;
