+++ pumpkin/src/fluid/fluids/water.rs
@@ -1,7 +1,7 @@
 use std::sync::Arc;
 
 use async_trait::async_trait;
-use pumpkin_data::fluid::{Falling, Fluid, FluidProperties, Level};
+use pumpkin_data::fluid::Fluid;
 use pumpkin_macros::pumpkin_fluid;
 use pumpkin_util::math::position::BlockPos;
 
@@ -13,7 +13,6 @@ use super::flowing_fluid::FlowingFluid;
 pub struct FlowingWater;
 
 const WATER_FLOW_SPEED: u16 = 5;
-type FlowingFluidProperties = pumpkin_data::fluid::FlowingWaterLikeFluidProperties;
 
 #[async_trait]
 impl PumpkinFluid for FlowingWater {
@@ -42,33 +41,6 @@ impl FlowingFluid for FlowingWater {
         1
     }
 
-    async fn get_source(&self, fluid: &Fluid, falling: bool) -> FlowingFluidProperties {
-        let mut source_props = FlowingFluidProperties::default(fluid);
-        source_props.level = Level::L8;
-        source_props.falling = if falling {
-            Falling::True
-        } else {
-            Falling::False
-        };
-        source_props
-    }
-
-    async fn get_flowing(
-        &self,
-        fluid: &Fluid,
-        level: Level,
-        falling: bool,
-    ) -> FlowingFluidProperties {
-        let mut flowing_props = FlowingFluidProperties::default(fluid);
-        flowing_props.level = level;
-        flowing_props.falling = if falling {
-            Falling::True
-        } else {
-            Falling::False
-        };
-        flowing_props
-    }
-
     async fn get_slope_find_distance(&self) -> i32 {
         4
     }
@@ -76,12 +48,5 @@ impl FlowingFluid for FlowingWater {
     async fn can_convert_to_source(&self, _world: &Arc<World>) -> bool {
         //TODO add game rule check for water conversion
         true
-    }
-
-    fn is_same_fluid(&self, fluid: &Fluid, other_state_id: u16) -> bool {
-        if let Some(other_fluid) = Fluid::from_state_id(other_state_id) {
-            return fluid.id == other_fluid.id;
-        }
-        false
     }
 }
